[Refactor/#28] divide messageDto(request, response)
