+++ src/main/java/com/messenger/chatty/global/util/CustomConverter.java
@@ -5,12 +5,9 @@ import com.messenger.chatty.domain.channel.entity.Channel;
 import com.messenger.chatty.domain.member.dto.response.MemberBriefDto;
 import com.messenger.chatty.domain.member.dto.response.MyProfileDto;
 import com.messenger.chatty.domain.member.entity.Member;
-import com.messenger.chatty.domain.message.dto.MessageDto;
+import com.messenger.chatty.domain.message.dto.response.MessageResponseDto;
 import com.messenger.chatty.domain.message.entity.Message;
 import com.messenger.chatty.domain.workspace.dto.response.WorkspaceBriefDto;
-import com.messenger.chatty.domain.workspace.dto.response.WorkspaceResponseDto;
-import com.messenger.chatty.domain.channel.entity.Channel;
-import com.messenger.chatty.domain.member.entity.Member;
 import com.messenger.chatty.domain.workspace.entity.Workspace;
 import org.springframework.data.domain.Page;
 
@@ -88,10 +85,10 @@ public class CustomConverter {
     }
 */
 
-    public static List<MessageDto> convertMessageResponse(Page<Message> messages) {
+    public static List<MessageResponseDto> convertMessageResponse(Page<Message> messages) {
          return messages.getContent().stream()
-                .map(message -> MessageDto.builder()
-                        .senderProfileImg(message.getId())
+                .map(message -> MessageResponseDto.builder()
+                        .senderProfileImg(message.getSenderProfileImg())
                         .senderUsername(message.getSenderUsername())
                         .senderNickname(message.getSenderNickname())
                         .regDate(TimeUtil.convertTimeTypeToLocalDateTime(message.getSendTime()))
